Add sysctl directives configuration tweaks
Some sysctl values (e.g. nfs ones) depend on some services existing in the system. Because of this, they can not be configured in the common role if the service is installed afterwards only on some nodes.
Also, the `group_sysctl_values` variables have been changed to `ovr_sysctl_values`, to clarify that the values in this variable override values already set in the common role.