Apply initial configuration through the 'user' user instead of root
The ssh daemon is configured to prohibit root logins using password authentication, which makes using the root user to perform the initial configuration a little cumbersome. Instead, the 'user' user is used, simply copying its public key to the servers beforehand.
However, the initial configuration playbook must be executed with the '-k' or '--ask-pass' parameter, and the 'sshpass' package must be installed in the control node.